Fix arbitrary code execution with subshell commands
If the shell would expand the contents of PS1 when drawing the prompt, a
raw ref name must not be included in PS1. This protects the user from
arbitrary code execution via specially crafted ref names.  For example,
a ref named 'refs/heads/$(IFS=_;cmd=sudo_rm_-rf_/;$cmd)' might cause the
shell to execute 'sudo rm -rf /' when the prompt is drawn.

https://github.com/git/git/blob/9d77b04/contrib/completion/git-prompt.sh#L324-L355